etl_hdfs/requirements.txt
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
docopt==0.6.2
hdfs==2.7.3
idna==3.11
influxdb-client==1.49.0
kafka==1.3.5
kafka-python==2.2.15
numpy==2.2.6
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
reactivex==4.1.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
